move printWinner logic into OutputView
