+++ README.md
@@ -20,5 +20,5 @@
 - [x] `Candidate` 대신 `Race`에 `List<Car>` 필드로 갖고 관리하기
 - [x] rename `RandomStrategy` → `MoveStrategy`
 - [x] `race`와 `round` 명칭 혼용 이슈 해결
-- [ ] 도메인의 `toString()`에 담긴 로직을 View로 이동하기
+- [x] 도메인의 `toString()`에 담긴 로직을 View로 이동하기
 - [ ] Exception 메세지를 View와 분리하기
