+++ src/main/java/game/view/dto/WinnerDto.java
@@ -3,11 +3,13 @@ package game.view.dto;
 import game.domain.Car;
 import game.domain.Winner;
 import lombok.Builder;
+import lombok.Getter;
 
 import java.util.List;
 
 public class WinnerDto {
 
+    @Getter
     private final List<Car> winners;
 
     @Builder
